Moved common-util to common, got rid of common-util, and added a new util module for really basic things that even ast depends on.
