--- common-util/src/main/java/com/graphicsfuzz/common/util/StreamGobblerLogger.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2018 The GraphicsFuzz Project Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.graphicsfuzz.common.util;
-
-import java.io.InputStream;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class StreamGobblerLogger extends StreamGobbler {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(StreamGobblerLogger.class);
-
-  private final String prefix;
-
-  public StreamGobblerLogger(InputStream inputStream, String prefix) {
-    super(inputStream);
-    this.prefix = prefix;
-  }
-
-  @Override
-  protected void handleLine(String line) {
-    LOGGER.info(prefix + line);
-  }
-
-  public StringBuffer getResult() {
-    return null;
-  }
-}
