+++ common/src/main/java/com/graphicsfuzz/common/util/IRandom.java
