+++ common/src/main/java/com/graphicsfuzz/common/util/Obfuscator.java
@@ -33,7 +33,9 @@ import com.graphicsfuzz.common.glslversion.ShadingLanguageVersion;
 import com.graphicsfuzz.common.typing.ScopeEntry;
 import com.graphicsfuzz.common.typing.ScopeTreeBuilder;
 import com.graphicsfuzz.common.typing.Typer;
-import com.graphicsfuzz.common.util.ExecHelper.RedirectType;
+import com.graphicsfuzz.util.ExecHelper;
+import com.graphicsfuzz.util.ExecHelper.RedirectType;
+import com.graphicsfuzz.util.ExecResult;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
