+++ fuzzerserver/src/main/java/com/graphicsfuzz/server/FuzzerServiceManagerImpl.java
@@ -16,10 +16,10 @@
 
 package com.graphicsfuzz.server;
 
-import com.graphicsfuzz.common.util.ExecHelper;
-import com.graphicsfuzz.common.util.ExecHelper.RedirectType;
-import com.graphicsfuzz.common.util.ExecResult;
-import com.graphicsfuzz.common.util.ToolPaths;
+import com.graphicsfuzz.util.ExecHelper;
+import com.graphicsfuzz.util.ExecHelper.RedirectType;
+import com.graphicsfuzz.util.ExecResult;
+import com.graphicsfuzz.util.ToolPaths;
 import com.graphicsfuzz.server.thrift.CommandInfo;
 import com.graphicsfuzz.server.thrift.CommandResult;
 import com.graphicsfuzz.server.thrift.FuzzerServiceManager;
