+++ parent-all/pom.xml
@@ -285,11 +285,6 @@ limitations under the License.
       </dependency>
       <dependency>
         <groupId>com.graphicsfuzz</groupId>
-        <artifactId>common-util</artifactId>
-        <version>1.0</version>
-      </dependency>
-      <dependency>
-        <groupId>com.graphicsfuzz</groupId>
         <artifactId>compare-asts</artifactId>
         <version>1.0</version>
       </dependency>
@@ -414,6 +409,12 @@ limitations under the License.
         <version>1.0</version>
       </dependency>
 
+      <dependency>
+        <groupId>com.graphicsfuzz</groupId>
+        <artifactId>util</artifactId>
+        <version>1.0</version>
+      </dependency>
+      
       <dependency>
         <groupId>com.graphicsfuzz</groupId>
         <artifactId>tool-webglfuzz</artifactId>
