+++ reducer/src/main/java/com/graphicsfuzz/reducer/tool/RandomFileJudge.java
@@ -16,10 +16,10 @@
 
 package com.graphicsfuzz.reducer.tool;
 
-import com.graphicsfuzz.common.util.ExecHelper.RedirectType;
-import com.graphicsfuzz.common.util.ExecResult;
+import com.graphicsfuzz.util.ExecHelper.RedirectType;
+import com.graphicsfuzz.util.ExecResult;
 import com.graphicsfuzz.common.util.IRandom;
-import com.graphicsfuzz.common.util.ToolHelper;
+import com.graphicsfuzz.util.ToolHelper;
 import com.graphicsfuzz.reducer.FileJudgeException;
 import com.graphicsfuzz.reducer.IFileJudge;
 import com.graphicsfuzz.reducer.util.ShaderJudgeUtil;
