+++ reducer/src/test/java/com/graphicsfuzz/reducer/reductionopportunities/InlineStructifiedFieldReductionOpportunitiesTest.java
@@ -22,11 +22,11 @@ import com.graphicsfuzz.common.ast.TranslationUnit;
 import com.graphicsfuzz.common.glslversion.ShadingLanguageVersion;
 import com.graphicsfuzz.common.tool.PrettyPrinterVisitor;
 import com.graphicsfuzz.common.transformreduce.Constants;
-import com.graphicsfuzz.common.util.ExecHelper.RedirectType;
-import com.graphicsfuzz.common.util.ExecResult;
+import com.graphicsfuzz.util.ExecHelper.RedirectType;
+import com.graphicsfuzz.util.ExecResult;
 import com.graphicsfuzz.common.util.Helper;
 import com.graphicsfuzz.common.util.ShaderKind;
-import com.graphicsfuzz.common.util.ToolHelper;
+import com.graphicsfuzz.util.ToolHelper;
 import com.graphicsfuzz.reducer.TestUtils;
 import java.io.File;
 import java.io.FileOutputStream;
