+++ util/src/main/java/com/graphicsfuzz/util/StreamGobbler.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.graphicsfuzz.common.util;
+package com.graphicsfuzz.util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
